dom/fetch/tests/test_bug1752761.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test bug1752761</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    add_task(async function () {
      const response = await fetch("multipart.sjs", {
        method: 'GET',
        mode: 'cors'
      });

      let p = response.formData().then(function (data) {
        SimpleTest.ok(data instanceof FormData, "Valid FormData extracted.");
      });
      await p;
    });
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
